add RPC interface to http clients
