+++ airframe-http/.jvm/src/main/scala/wvlet/airframe/http/client/JavaAsyncClient.scala
@@ -16,7 +16,7 @@ package wvlet.airframe.http.client
 import wvlet.airframe.http.HttpClientConfig
 import wvlet.airframe.http.HttpMessage.{Request, Response}
 
-import scala.concurrent.Future
+import scala.concurrent.{ExecutionContext, Future}
 
 /**
   * An wrapper of JavaHttpSyncClient for supporting async response
@@ -24,7 +24,8 @@ import scala.concurrent.Future
   */
 class JavaAsyncClient(syncClient: JavaHttpSyncClient) extends AsyncClient {
 
-  private[http] def config: HttpClientConfig = syncClient.config
+  private[http] def config: HttpClientConfig           = syncClient.config
+  private[http] val executionContext: ExecutionContext = syncClient.executionContext
 
   override def send(req: Request, requestFilter: Request => Request): Future[Response] = {
     syncClient.sendAsync(req, requestFilter)
