+++ airframe-http/.jvm/src/main/scala/wvlet/airframe/http/client/JavaHttpSyncClient.scala
@@ -39,13 +39,13 @@ import scala.util.control.NonFatal
 class JavaHttpSyncClient(serverAddress: ServerAddress, private[client] val config: HttpClientConfig)
     extends client.SyncClient {
 
-  private val javaHttpClient: HttpClient     = newClient(config)
-  private val circuitBreaker: CircuitBreaker = config.circuitBreaker.withName(s"${serverAddress}")
-  private implicit val ec: ExecutionContext  = config.newExecutionContext
+  private val javaHttpClient: HttpClient                          = newClient(config)
+  private val circuitBreaker: CircuitBreaker                      = config.circuitBreaker.withName(s"${serverAddress}")
+  private[client] implicit val executionContext: ExecutionContext = config.newExecutionContext
 
   override def close(): Unit = {
     // It seems Java Http Client has no close method
-    ec match {
+    executionContext match {
       case e: ExecutorService =>
         // Close the thread pool
         e.shutdownNow()
@@ -60,7 +60,7 @@ class JavaHttpSyncClient(serverAddress: ServerAddress, private[client] val confi
       .connectTimeout(java.time.Duration.ofMillis(config.connectTimeout.toMillis))
       // Wrap Scala's ExecutionContext as Java's Executor
       .executor(new Executor {
-        override def execute(command: Runnable): Unit = ec.execute(command)
+        override def execute(command: Runnable): Unit = executionContext.execute(command)
       })
       .build()
   }
